docs and file renaming
