--- uflx/expression.py
@@ -1,7 +1,0 @@
-"""Expression."""
-
-from abc import ABC
-
-
-class AbstractExpression(ABC):
-    """Abstract base class for expressions."""
