+++ uflx/expressions.py
@@ -1,0 +1,15 @@
+# Copyright (C) 2025 Matthew Scroggs and Garth N. Wells
+#
+# This file is part of UFLx (https://www.fenicsproject.org)
+#
+# SPDX-License-Identifier:    MIT
+"""Expression.
+
+An expression is...
+"""
+
+from abc import ABC
+
+
+class AbstractExpression(ABC):
+    """Abstract base class for expressions."""
