+++ uflx/integrals.py
@@ -1,8 +1,13 @@
-"""Integral."""
+# Copyright (C) 2025 Matthew Scroggs and Garth N. Wells
+#
+# This file is part of UFLx (https://www.fenicsproject.org)
+#
+# SPDX-License-Identifier:    MIT
+"""Integrals."""
 
 from abc import ABC, abstractmethod
 
-from uflx.expression import AbstractExpression
+from uflx.expressions import AbstractExpression
 
 
 class AbstractMeasure(ABC):
