+++ uflx/operators.py
@@ -1,9 +1,14 @@
+# Copyright (C) 2025 Matthew Scroggs and Garth N. Wells
+#
+# This file is part of UFLx (https://www.fenicsproject.org)
+#
+# SPDX-License-Identifier:    MIT
 """Operators."""
 
 from abc import abstractmethod
 
-from uflx.expression import AbstractExpression
-from uflx.function import AbstractFunction
+from uflx.expressions import AbstractExpression
+from uflx.functions import AbstractFunction
 
 
 class AbstractOperator(AbstractExpression):
